fix(spectralquant): preserve indices dtype in 2/4-bit pack padding
Address claude-review feedback on #257:

- `pack_indices` was hardcoding padding zeros to `mx.uint8`, which would
  silently upcast when `indices` is int32 (the real caller path from
  `spectral_quantize` → `mx.argmin`). Match the concat dtype to the input
  so we don't lean on `_tq_pack`'s implicit `.astype(mx.uint8)`.
- Parametrize `TestPackUnpack` over `[np.uint8, np.int32]` so the
  dtype-promotion path the real code hits is actually exercised.

+++ olmlx/engine/spectralquant.py
@@ -44,7 +44,7 @@ def pack_indices(indices: mx.array, bits: int) -> mx.array:
     pad = (-dim) % factor
     if pad:
         indices = mx.concatenate(
-            [indices, mx.zeros(indices.shape[:-1] + (pad,), dtype=mx.uint8)],
+            [indices, mx.zeros(indices.shape[:-1] + (pad,), dtype=indices.dtype)],
             axis=-1,
         )
     return _tq_pack(indices, bits)
